requirements.txt
pandas>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
